+++ angular/lib/src/compiler/ast_directive_normalizer.dart
@@ -7,6 +7,7 @@ import 'package:angular_compiler/cli.dart';
 
 import 'compile_metadata.dart';
 import 'directive_normalizer.dart';
+import 'expression_parser/visitor.dart';
 import 'style_url_resolver.dart' show extractStyleUrls, isStyleUrlResolvable;
 
 /// Loads the content of `templateUrl` and `styleUrls` to normalize directives.
@@ -17,8 +18,14 @@ import 'style_url_resolver.dart' show extractStyleUrls, isStyleUrlResolvable;
 /// The normalizer also resolves inline style and stylesheets in the template.
 class AstDirectiveNormalizer implements DirectiveNormalizer {
   final NgAssetReader _reader;
-
   const AstDirectiveNormalizer(this._reader);
+
+  void _parseExpressionsWithLegacyParser(
+    ast.TemplateAst astNode,
+    List<CompileIdentifierMetadata> exports,
+  ) {
+    astNode.accept(new LegacyExpressionVisitor(exports: exports));
+  }
 
   @override
   Future<CompileDirectiveMetadata> normalizeDirective(
@@ -28,7 +35,11 @@ class AstDirectiveNormalizer implements DirectiveNormalizer {
     if (!directive.isComponent) {
       return new Future.value(directive);
     }
-    return normalizeTemplate(directive.type, directive.template).then((result) {
+    return normalizeTemplate(
+      directive.type,
+      directive.template,
+      exports: directive.exports,
+    ).then((result) {
       return new CompileDirectiveMetadata(
         type: directive.type,
         metadataType: directive.metadataType,
@@ -57,8 +68,9 @@ class AstDirectiveNormalizer implements DirectiveNormalizer {
   @override
   Future<CompileTemplateMetadata> normalizeTemplate(
     CompileTypeMetadata directiveType,
-    CompileTemplateMetadata template,
-  ) async {
+    CompileTemplateMetadata template, {
+    List<CompileIdentifierMetadata> exports,
+  }) async {
     template ??= new CompileTemplateMetadata(template: '');
     if (template.template != null) {
       await _validateTemplateUrlNotMeant(template.template, directiveType);
@@ -68,6 +80,7 @@ class AstDirectiveNormalizer implements DirectiveNormalizer {
         template.template,
         directiveType.moduleUrl,
         template.preserveWhitespace,
+        exports: exports,
       );
     }
     if (template.templateUrl != null) {
@@ -82,6 +95,7 @@ class AstDirectiveNormalizer implements DirectiveNormalizer {
           templateContent,
           sourceAbsoluteUrl,
           template.preserveWhitespace,
+          exports: exports,
         );
       });
     }
@@ -118,8 +132,9 @@ class AstDirectiveNormalizer implements DirectiveNormalizer {
     CompileTemplateMetadata templateMeta,
     String template,
     String templateAbsUrl,
-    bool preserveWhitespace,
-  ) {
+    bool preserveWhitespace, {
+    List<CompileIdentifierMetadata> exports: const [],
+  }) {
     // Parse the template, and visit to find <style>/<link> and <ng-content>.
     final visitor = new _TemplateNormalizerVisitor();
     final parsedNodes = ast.parse(
@@ -128,7 +143,11 @@ class AstDirectiveNormalizer implements DirectiveNormalizer {
       // Otherwise, the analyzer crashes today seeing an 'asset:...' URL.
       sourceUrl: Uri.parse(templateAbsUrl).replace(scheme: 'file').toFilePath(),
     );
-    parsedNodes.forEach((a) => a.accept(visitor));
+
+    for (final node in parsedNodes) {
+      node.accept(visitor);
+      _parseExpressionsWithLegacyParser(node, exports);
+    }
 
     final allInlineStyles = templateMeta.styles + visitor.inlineStyles;
     final allExternalStyles = <String>[];